Fixes about huge payload receiving
So, remember I told "payload size that is near to 2GB"?
We need to talk about that. Sadly,there is no such thing. The reason is...
You found it. Memory.
As you see from code, we define payloadData and assign those payload chunks that is sent by our client. And as you know, RAM and Storage in Raspberry Pico is limited. And we can't go over it.

But no worries, big payloads are still there. I, personally tested a payload that is size of 10Kb and it worked very well. My point is, payload will work till the memory need for payload is not over limits. The limits include available memory and some memory that is required to execute given payload.

In the end, payloads can be big. But not too big.

Alternatively, if you need payload that is over the Pico's memory(And for what you are gonna use that huge payload...), you can do this:
1) Buy an sd card
2) Buy and sd card module for pico
3) Plug it in
4) Receive payload in chunks and then write it into the file in sd card with chunks
5) And when receiving is done, open the file where you saved huge payload and read it line by line. And then run each line individually.

That is it. That is the summary for this update.
Happy ~~hacking!~~ joking with your friends!