Added configuration support
Yes, now from this era, we are not only dependent on existing networks to connect and then make our way through. NOW WE CAN CREATE OUR OWN NETWORK!
RISE MY SOLDIERS! TODAY,WE ARE MAKING OUR SECRET LINE TO DO WHATEVER WE WANT!
TODAY IS THE DAY,WHICH WE ARE THE RULERS OF THE WORLDS!
LONG LIVE PICO!!!

*Pls, Hollywood, hire me as writer*
Anyways ;))
Jokes aside, we actually can make our AP now in case we don't have any known networks to operate on. Prefer you to check out the Usage section in ReadMe.
Happy hacking! ;))